backend/requirements.txt
alembic==1.13.1
APScheduler==3.10.4
blinker==1.7.0
click==8.1.7
exceptiongroup==1.1.3
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.2
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.8.0
tzlocal==5.2
Werkzeug==3.0.1
alembic==1.13.1
APScheduler==3.10.4
blinker==1.7.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
Werkzeug==3.0.1
alembic==1.13.1
APScheduler==3.10.4
blinker==1.7.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.1
gunicorn==22.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
Werkzeug==3.0.1
